Client.Assets.Generic

Documentation

newtype AssetID a

A unique identifier for assets. The identifier has a phantom type which ensures that id's of different assets can not be mixed.

Constructors

AssetID

Fields

unAssetID :: Text

data AssetLoader a b c

Constructors

AssetLoader

Fields

store :: BasicHashTable Text a

extData :: b

load :: c -> b -> FilePath -> IO (Either Text [(AssetID a, a)])

delete :: a -> IO ()

addAsset :: MonadIO m => AssetLoader a b c -> a -> AssetID a -> m ()

Utiltiy function to manually add an asset

getAsset :: (MonadIO m, MonadError Text m) => AssetLoader a b c -> AssetID a -> m a

Utility function to retrieve a single asset

getAllIDs :: MonadIO m => AssetLoader a b c -> m [AssetID a]

Utility function for retrieving all assets

loadAsset :: (MonadIO m, MonadError Text m) => AssetLoader a b c -> c -> FilePath -> m [AssetID a]

Utility function to load a single asset

loadAsset_ :: (MonadIO m, MonadError Text m) => AssetLoader a b () -> FilePath -> m [AssetID a]

Utility function to load a single asset for the special case of `c == ()`

loadAssets

Arguments

:: (MonadError Text m, MonadIO m)

=> AssetLoader a b c
AssetLoader to load textures into

-> [(c, FilePath)]
Paths to asset files. Note that these should be specified relative to the root of the git directory

-> (Float -> Text -> m ())
Callback function. This function will be called whenever a new asset is loading with the percentage of loaded files and the name of the currently loading asset. If this is not needed, set it to `return ()`.

-> m ()

Utility function to load a bunch of assets into an AssetLoader with a callback function. Also normalizes the paths via the cabal paths.

loadAssets_ :: (MonadIO m, MonadError Text m) => AssetLoader a b () -> [FilePath] -> (Float -> Text -> m ()) -> m ()

Same as loadAssets specialized for `c == ()`

deleteAssets :: MonadIO m => AssetLoader a b c -> m ()

Utility function to delete allassets in an AssetLoader