Client.States.MenuState.Loop

Description

Documentation

loop :: (MonadIO m, MonadState State m, MonadReader Resources m) => CommonState -> m ()

The loop while being in the menu. Note that loop should _not_ pass any errors further along, unless they are fatal. There is no recovery beyond this point, hence no MonadError.