Client.Types

Documentation

data AppState

The AppState holds all information of the entire application.

Constructors

AppState

Fields

_running :: Bool

_game :: GameState

_sprites :: SpriteLoaderType

_frameState :: FrameState

running :: Lens' AppState Bool

game :: Lens' AppState GameState

sprites :: Lens' AppState SpriteLoaderType

frameState :: Lens' AppState FrameState

type Game a = ReaderT Resources (StateT AppState IO) a

data FrameState

The frame state contains information on the app's fps.

Constructors

FrameState

Fields

_fpsCounter :: FpsCounter

_frameBegin :: UTCTime

_dt :: DeltaTime

fpsCounter :: Lens' FrameState FpsCounter

frameBegin :: Lens' FrameState UTCTime

data NetworkState

The network state of a client. This data type contains every information that is available for a connection to a game server.

Constructors

NetworkState

Fields

_lastPacket :: UTCTime

_lastID :: MessageCount

_ackField :: AckField

_socket :: Socket

_tbqueue :: TBQueue ServerToClientMsg

lastPacket :: Lens' NetworkState UTCTime

lastID :: Lens' NetworkState MessageCount

ackField :: Lens' NetworkState AckField

socket :: Lens' NetworkState Socket

tbqueue :: Lens' NetworkState (TBQueue ServerToClientMsg)

data GameState

GameState has the information of the current state of the game. In can be either a ingame state, or a menu state.

Constructors

InGameState

Fields

_physicalObjects :: PhysicalObjects

_accumTime :: DeltaTime

Accumulated time for fixed physics step

_gameObjects :: IntMap GameObject

_controls :: Controls

_netState :: NetworkState

_keymap :: KeyMap

MenuState

physicalObjects :: Traversal' GameState PhysicalObjects

dt :: Lens' FrameState DeltaTime

accumTime :: Traversal' GameState DeltaTime

gameObjects :: Traversal' GameState (IntMap GameObject)

controls :: Traversal' GameState Controls

netState :: Traversal' GameState NetworkState

keymap :: Traversal' GameState KeyMap

data Resources

Resources are things that do not change during a complete app cycle.

Constructors

Resources

Fields

_window :: Window

_renderer :: Renderer

window :: Lens' Resources Window

renderer :: Lens' Resources Renderer

data FpsCounter

Constructors

FpsCounter

Fields

maxFrames :: Int

fpsL :: [FlType]