Client.Video.Sprite

Documentation

renderSprite :: (MonadIO m, MonadError Text m) => Renderer -> SpriteLoaderType -> SpriteID -> Maybe (Rectangle CInt) -> CDouble -> V2 Bool -> m ()

Render a sprite

TODO: The renderer as an argument is unnecessary if we take it out of the Textureloader

noFlip :: V2 Bool

Draw sprite without flipping

renderGameObject :: (MonadIO m, MonadError Text m) => Renderer -> SpriteLoaderType -> Resolution -> GameObject -> m ()

updateRenderInfo :: PhysicalObjects -> GameObject -> GameObject

Update the sprite position with the position of the physical Object

relToAbs :: Resolution -> Rectangle FlType -> Rectangle CInt