PurpleMuon.Input.Types

Description

Purple Moun Clients have `8` possible keys they can press. These are

accelerate (up key)

turn left (left key)

turn right (right key)

decelerate (down key)

fire 1 (space)

fire 2 (s)

fire 3 (d)

fire 4 (f)

These controls are send from the clients to the server 30 times a second. They packed into a Word8, where the bits signalize pressed (1) or not pressed (0) in the order mentioned above.

Documentation

data Controls

A data type that stores information about the current keyboard state

Constructors

Controls

Fields

accelerate :: Bool

turn_left :: Bool

turn_right :: Bool

decelerate :: Bool

fire1 :: Bool

fire2 :: Bool

fire3 :: Bool

fire4 :: Bool

data KeyMap

A data type that stores a keymap. A keymap is a mapping from SDL scancodes to controls defined in the game

Constructors

KeyMap

Fields

km_accel :: Scancode

km_turn_left :: Scancode

km_turn_right :: Scancode

km_decel :: Scancode

km_fire1 :: Scancode

km_fire2 :: Scancode

km_fire3 :: Scancode

km_fire4 :: Scancode

packControls :: Controls -> Word8

unpackControls :: Word8 -> Controls