PurpleMuon.Network.Types

Description

Documentation

newtype RawMessage

A binary message that is send over the network

Constructors

RawMessage

Fields

unRawMessage :: ByteString

newtype NakedMessage

A binary message without a uuid header

Constructors

NakedMessage

Fields

unNakedMessage :: ByteString

newtype UUID

A uuid is a header that the server and client prepends to every message. A message that does not have such a header is discarded immediately

Constructors

UUID

Fields

unUUID :: ByteString

newtype Payload

The payload of a Message

Constructors

Payload

Fields

unPayload :: ByteString

newtype MessageCount

A counter to count both the remote and local messages

Constructors

MessageCount

Fields

unMessageCounter :: Word32

newtype AckField

A field indicating which of the last 32 messages have arrived.

Constructors

AckField

Fields

unAckField :: Word32

newtype Offset

A difference between to Message counts

Constructors

Offset

Fields

unOffset :: Int

data NetworkConfig

A Configuration of a connection

Constructors

ConnectionConfig

Fields

maxClients :: Int

data NetworkState

Constructors

NetworkState

Fields

socket :: Socket

data ConnectionState

Constructors

ConnectionState

Fields

addr :: SockAddr

latestCounter :: MessageCount

data ServerToClientMsg

Messages the Server sends to the Clients

Constructors

Ping
A simple ping package to determine network latency

Update PhysicalObjects
An update the server sends to the clients

data ClientToServerMsg

Messages the Clients send to the Server

Constructors

RequestConnection PlayerName
Request a connection to the server. The sever will answer with ping commands to establish the network latency if the request is granted

Pong MessageCount
Answer to a ping command. The MessageCount indicates which

newtype PlayerName

A player name

Constructors

PlayerName

Fields

unPlayerName :: Text