PurpleMuon.Physics.Constants

Description

Documentation

g :: GravitationalConstant

physicalSize :: PhysicalSize

physicsStep :: DeltaTime

Do physics at 30 Hz

minimumDistance :: FlType

The minimum distance two objects are allowed to have This makes the integration numerically much more stable