PurpleMuon.Physics.Types

Description

Basic types

type PhyObjKey = Key PhysicalObject

A physical object identifier

newtype Mass

The mass of an object

Constructors

Mass

Fields

unMass :: FlType

newtype Velocity

The velocity of an object

Constructors

Velocity

Fields

unVelocity :: V2 FlType

newtype Acceleration

The acceleration of an object

Constructors

Acceleration

Fields

unAcceleration :: V2 FlType

newtype DeltaTime

A time delta Meassured in seconds

Constructors

DeltaTime

Fields

unDeltaTime :: FlType

newtype Force

A Force

Constructors

Force

Fields

unForce :: V2 FlType

newtype GravitationalConstant

The gravitaional constant

Constructors

GravitationalConstant

Fields

unGravitationalConstant :: FlType

Combined types

data PhysicalObject

A physical object

Constructors

PhysicalObject

Fields

_mass :: Mass

The mass of the object

_pos :: Position

The position of the object

_vel :: Velocity

The velocity of the object

_static :: ObjType

Is this object moving?

_gravitating :: ObjGrav

Is this object gravitating?

mass :: Lens' PhysicalObject Mass

pos :: Lens' PhysicalObject Position

vel :: Lens' PhysicalObject Velocity

static :: Lens' PhysicalObject ObjType

gravitating :: Lens' PhysicalObject ObjGrav

newtype PhysicalSize

The physical size of the playing board

Constructors

PhysicalSize

Fields

unPhysicalSize :: V2 FlType

newtype Derivative

The derivaties of a physical object numerical integrater

Constructors

Derivative

Fields

unDerviative :: (Velocity, Force)

IntMap types

type Derivatives = IntMap Derivative

A collection of derivaties, indexed by the uuid of the objects

type PhysicalObjects = IntMap PhysicalObject

A collection of objects, indexed by the uuid of the objects

type Forces = IntMap Force

A collection of forces, indexed by the uuid of the objects

Boolean types

data ObjGrav

The type of a gravitating body

Constructors

Gravitating
the object exerts gravitational forces

NonGravitating
the object only recerives grav. forces

data ObjType

The type of a physical object

Constructors

Static
the objects position is fixed

NonStatic
the objects position is not fixed