PurpleMuon.Types

Description

This module defined types that are used in multiple subcomponents. Instead of having different versions in each subcompoenent, they are collected here. As such, this module is kind of a collection of things without a common denominator.

Documentation

type FlType = Float

The floating point type used in PurpleMuon

newtype Position

A position type describing a position on the screen. Screen positions are always int the range `[0,1] x [0,1]` and will be rescaled when drawn according to the screen resolution. The physics engine uses the same positions to describe the physical objects.

Constructors

Position

Fields

unPosition :: V2 FlType