PurpleMuon.Util.Frames

Description

This module defines helper function for frame management. It allows to define a wanted frame length and wait if the frame was shorter. It also abstracts over storing frame lengths and calculating FPS.

Documentation

frameBegin

Arguments

:: MonadIO m

=> (UTCTime -> m ())
function to store the time

-> m ()

Call at the beginning of a frame. Updates the frameBegin timer.

manageFps

Arguments

:: MonadIO m

=> NominalDiffTime
the desired frame time

-> m UTCTime
get the frame begin

-> (FlType -> m ())
store dt

-> m ()

Handle frame time. The function will then calculate the remaining time, sleep the thread, update the frame counter and the dt variable of the game state