PurpleMuon.Util.MonadError

Description

This module defines common lifting operations into MonadError.

Documentation

liftMaybe

Arguments

:: MonadError e m

=> e
Error to be thrown if Nothing

-> Maybe a
Value that has to be Just

-> m a

Lift a Maybe into an error monad

liftEitherWith

Arguments

:: MonadError e m

=> (f -> e)
Transformation for left value

-> Either f a
Value that has to be Right

-> m a

Lift Either into error monad This version additionally allows to specify a transformation for the Left value.

liftEither :: MonadError e m => Either e a -> m a

Lift Either into error monad Throws error on Left Value.