Server.Types

Documentation

data WaitingState

The state of a server waiting for connections

Constructors

WaitingState

Fields

_clientsConnected :: [ClientConnection]

data GameState

The state of a server in game

Constructors

GameState

Fields

_gObjs :: IntMap GameObject

_pObjs :: PhysicalObjects

_frameBegin :: UTCTime

_clients :: [ClientConnection]

_intStep :: Word32

Count of integration steps

gObjs :: Lens' GameState (IntMap GameObject)

pObjs :: Lens' GameState PhysicalObjects

frameBegin :: Lens' GameState UTCTime

clients :: Lens' GameState [ClientConnection]

logger :: Lens' Resources LoggerSet

intStep :: Lens' GameState Word32

type WaitingServer a = ReaderT Resources (StateT WaitingState IO) a

clientsConnected :: Iso' WaitingState [ClientConnection]

type Server a = ReaderT Resources (StateT GameState IO) a

data Resources

Read only resources that the server has access to

Constructors

Resources

Fields

_tbqueue :: TBQueue NakedMessage

_socket :: Socket

_logger :: LoggerSet

_uuid :: ProtocolUUID

tbqueue :: Lens' Resources (TBQueue NakedMessage)

socket :: Lens' Resources Socket

uuid :: Lens' Resources ProtocolUUID

data ClientConnection

A client connection saves all the data the server knows about a client

Constructors

ClientConnection

Fields

_addr :: SockAddr

_name :: Text

_gameObj :: GameObjKey

addr :: Lens' ClientConnection SockAddr

name :: Lens' ClientConnection Text

gameObj :: Lens' ClientConnection GameObjKey