Client.Assets.Generic

Documentation

newtype AssetID a

A unique identifier for assets. The identifier has a phantom type which ensures that id's of different assets can not be mixed.

Constructors

AssetID

Fields

unAssetID :: Text

class AssetLoader al where

A AssetLoader is responsible for managing assets. It can load assets and store them for easy retrival.

Minimal complete definition

loadAsset, getAsset, deleteAsset, getAllIDs

Associated Types

data Asset al :: *

Type of an asset

Methods

loadAsset :: (MonadIO m, MonadError Text m) => al -> FilePath -> m [AssetID (Asset al)]

Load an asset (maybe multiple assets) into an AssetLoader. Returns the AssetIDs of the loaded types.

WARNING: loadAsset will always normalize the given FilePath via the cabal paths package. Hence all paths should be specified relative to the root of the git repository. Also make sure that the files are mentioned in the package.yaml file under data-files.

getAsset :: (MonadIO m, MonadError Text m) => al -> AssetID (Asset al) -> m (Asset al)

Retreive asset from AssetLoader

deleteAsset :: MonadIO m => al -> AssetID (Asset al) -> m ()

Delete an asset

getAllIDs :: MonadIO m => al -> m [AssetID (Asset al)]

Get all AssetIDs

data HashmapLoader a ext

A simple implementation of an AssetLoader via Hashmaps a is the type of the Asset to be loaded and stored and ext can be any extra data that needs to be used when loading an asset from a file.

Constructors

HashmapLoader

Fields

store :: BasicHashTable Text a

extraData :: ext

loadFromFile :: ext -> FilePath -> IO (Either Text [(AssetID a, a)])

delete :: ext -> a -> IO ()

loadAssets

Arguments

:: (MonadError Text m, MonadIO m, AssetLoader a)

=> a
AssetLoader to load textures into

-> [FilePath]
Paths to asset files. Note that these should be specified relative to the root of the git directory

-> (Float -> Text -> m ())
Callback function. This function will be called whenever a new asset is loading with the percentage of loaded files and the name of the currently loading asset. If this is not needed, set it to `return ()`.

-> m ()

Utility function to load a bunch of assets into an AssetLoader with a callback function. Also normalizes the paths via the cabal paths.

deleteAssets :: (AssetLoader a, MonadIO m) => a -> m ()

Utility function to delete allassets in an AssetLoader